+++ src/main/java/com/devbliss/gpullr/controller/dto/notification/NotificationDto.java
@@ -1,22 +1,9 @@
 package com.devbliss.gpullr.controller.dto.notification;
 
-import com.devbliss.gpullr.domain.notifications.NotificationType;
+import java.util.List;
 
-/**
- * DTO {@link NotificationDto} to be transferred by the {@link com.devbliss.gpullr.controller.NotificationController}
- * to the calling client.
- */
 public class NotificationDto {
+  public List<SystemNotificationDto> systemNotifications;
 
-  public long id;
-
-  public String createdAt;
-
-  public NotificationType type;
-
-  public String actorName;
-
-  public String repoTitle;
-
-  public String pullRequestTitle;
+  public List<UserNotificationDto> userNotifications;
 }