+++ src/main/java/com/devbliss/gpullr/domain/notifications/UserNotificationType.java
@@ -1,0 +1,9 @@
+package com.devbliss.gpullr.domain.notifications;
+
+/**
+ * Notification type used in the {@link com.devbliss.gpullr.service.UserNotificationService} and
+ * the {@link UserNotification}.
+ */
+public enum UserNotificationType {
+  PULLREQUEST_CLOSED
+}